refactor(XZMocoa): 更新宏实现及文档，优化绑定机制与角色推断
- 重命名宏文件，调整宏类型声明和注册方式，提高代码一致性
- 规范 @mocoa 宏角色注释，明确属性和方法的 @objc 修饰规则
- 修正 @key 宏只添加 @objc，取消 dynamic，增强 KVO 兼容性
- 优化 @key 宏对已有 set/didSet 的检测，改为检测 sendActions 调用
- 扩展 @bind 宏，新增 @link 单次绑定支持，完善绑定声明族与用法
- 细化 View 中绑定属性与方法的多样绑定形式和自动推断逻辑
- 增加对可选属性绑定时 didSet 生成及动态监听警告提示
- 完善 #module 宏参数校验规则，增强编译期参数合法性检测
- 统一错误和警告提示语，区分不同宏及角色使用场景
- 修订示例代码，演示 @mocoa、@key、@bind、@link 宏协同应用
- 修正文档说明及代码注释，提升阅读理解和使用准确性
- 修复拼写和表述错误，调整内部注释及示例格式
- 更新 ObjC 代码注释与方法实现细节，确保语义清晰准确
- 文档中新增 @link 宏介绍，突出单次绑定与持续绑定的区别

+++ Sources/ObjC/XZMocoa/XZMocoaView/ViewModel/XZMocoaViewModel.m
@@ -22,8 +22,7 @@
 }
 
 - (void)dealloc {
-    // 移除 kvo
-    // [_observer removeAllTargets];
+    // 移除对数据模型的 KVO 观察（见下方 _removeModelObserverIfNeeded:）。
     
     // 不能像下面这样使用 for-in 语句。
     // for (XZMocoaViewModel *viewModel in subViewModels) {
@@ -111,7 +110,7 @@
 }
 
 - (NSString *)description {
-    return [NSString stringWithFormat:@"<%@: %p, isReady = %@; subViewModels = (%ld objects)>", self.class, self, @(self.isReady), self.subViewModels.count];
+    return [NSString stringWithFormat:@"<%@: %p, isReady = %@; subViewModels = (%lu objects)>", self.class, self, @(self.isReady), (unsigned long)self.subViewModels.count];
 }
 
 - (void)setModel:(id)model {
